[ANDROAPP-3651] Refactor EventCreationType
