+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventInitial/EventInitialPresenter.java
@@ -18,7 +18,7 @@ import org.dhis2.usescases.eventsWithoutRegistration.eventCapture.EventFieldMapp
 import org.dhis2.utils.D2EditionMapper;
 import org.dhis2.utils.DateUtils;
 import org.dhis2.utils.DhisTextUtils;
-import org.dhis2.utils.EventCreationType;
+import org.dhis2.commons.data.EventCreationType;
 import org.dhis2.utils.Result;
 import org.dhis2.form.data.RulesUtilsProvider;
 import org.dhis2.utils.analytics.AnalyticsHelper;
