+++ app/src/main/java/org/dhis2/usescases/programStageSelection/ProgramStageSelectionRepositoryImpl.kt
@@ -6,7 +6,7 @@ import io.reactivex.functions.Function5
 import java.util.ArrayList
 import org.dhis2.Bindings.toRuleAttributeValue
 import org.dhis2.form.data.RulesRepository
-import org.dhis2.utils.EventCreationType
+import org.dhis2.commons.data.EventCreationType
 import org.dhis2.utils.Result
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.enrollment.Enrollment
