+++ commons/src/main/java/org/dhis2/commons/data/EventCreationType.java
@@ -1,4 +1,4 @@
-package org.dhis2.utils;
+package org.dhis2.commons.data;
 
 /**
  * QUADRAM. Created by ppajuelo on 10/10/2018.
